Balance input dataset
I had over 100:1 ratio between the two different things I was trying to
detect. The result was that while the error was small, it was still
larger than the thing that it was trying to detect, so it was very
possible that the rare results were entirely missed and therefore
included in the error.

Now the two states are much more closely matched:

-rw-r--r-- 1 ksandom users 9.7M Jan 11 14:44 ./data/samples/10/combinedBinaryA.csv
   6467 0
   9219 1
-rw-r--r-- 1 ksandom users 2.5M Jan 11 14:44 ./data/samples/10/combinedBinaryB.csv
   1541 0
   2460 1
-rw-r--r-- 1 ksandom users 64M Jan 11 14:44 ./data/samples/10/combinedBinaryC.csv
 103829 1